General/global: Another set of changes towards better LLP64 compatibility (besides existing ILP32 and LP64 compatibility).
git-svn-id: https://srv7.svn-repos.de/dev123/projects/cafu/trunk@100 5fe0b3f3-48df-4727-b460-e6dda9226989

+++ Libs/Models/Model_mdl.cpp
@@ -259,7 +259,7 @@ ModelMdlT::ModelMdlT(const std::string& FileName_) /*throw (ModelT::LoadError)*/
         for (unsigned long i=0; i<MaterialName.length(); i++) if (MaterialName.at(i)=='\\') MaterialName.at(i)='/';
 
         // Strip any extension.
-        for (unsigned long i=MaterialName.length(); i>0; i--)
+        for (size_t i=MaterialName.length(); i>0; i--)
         {
             if (MaterialName.at(i-1)=='/') break;
